work in progress caching
- tidy
- delete inmemmeetings service
- delete stub
- have uniform names for tests
