+++ src/config/runtime/unit.ts
@@ -1,10 +1,8 @@
 import {EnvironmentConfig} from '../../model/EnvironmentConfig';
-
-import {MSGraphTokenProvider} from '../../services/tokens/MSGraphTokenProvider';
+import {RuntimeConfig} from '../../model/RuntimeConfig';
 
 import {MockUserService} from '../../services/users/MockUserService';
 
-import {RuntimeConfig} from '../../model/RuntimeConfig';
 import {generateUnitRoomLists} from '../bootstrap/rooms';
 import {MockRoomService} from '../../services/rooms/MockRoomService';
 import {MockJWTTokenProvider} from '../../services/tokens/MockJWTTokenProvider';
@@ -15,7 +13,6 @@ import {MockGraphTokenProvider} from '../../services/tokens/MockGraphTokenOperat
 
 export function provideUnitRuntime(environment: EnvironmentConfig): RuntimeConfig {
   const jwtTokenProvider = new MockJWTTokenProvider(environment.jwtTokenSecret);
-  const graphAPIParameters = environment.graphAPIParameters;
 
   return new RuntimeConfig(environment.port,
                            new MockPasswordStore(),
