--- src/services/stub/README.md
@@ -1,1 +1,0 @@
-Things in this directory should be purely stubs and provide stubbed or canned data.
