--- test/services/meetings/InmemMeetings-spec.ts
@@ -1,6 +1,0 @@
-// import {InmemMeetingService} from '../../../src/services/meetings/InmemMeetingService';
-// import StatefulSpec from '../Stateful-spec';
-//
-// const svc = new InmemMeetingService();
-//
-// StatefulSpec(svc, 'In-memory meetings services');
